+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- Improve handling of organization-related images to prevent unwanted cropping
+  of organization logos.
+
 ## [1.3.0] - 2019-06-20
 
 ### Changed